requirements.txt
absl-py==0.7.1
args==0.1.0
astor==0.7.1
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
Click==7.0
clint==0.5.1
Flask==1.0.2
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.8
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lazy-object-proxy==1.3.1
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
numpy==1.16.3
pbr==5.2.0
Pillow==6.0.0
protobuf==3.7.1
pylint==2.3.1
PyYAML==5.1
requests==2.21.0
rope==0.14.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.3.5
urllib3==1.24.2
Werkzeug==0.15.2
wrapt==1.11.1
